refactor: improve Google Cast cmd handling, clean up setup flow (#73)
* refactor: fallback volume control, Cast command status codes

For volume up, down and mute use normal Android TV volume control
if Google Cast volume control is enabled but not available.

Return SERVICE_UNAVAILABLE if Google Cast is required for a command,
like seek, but not connected.

Return SERVER_ERROR if a Google Cast command failed.

* refactor: de-duplicate setup-flow configuration options

Only define configuration settings once and re-use for initial device
setup and for device reconfiguration.
Add missing German texts.

* refactor: Pre-select configure action in setup flow if a device exists

* chore: missing German text

+++ src/config.py
@@ -40,7 +40,7 @@ class AtvDevice:
     """Enable Chromecast features."""
     use_chromecast_volume: bool = False
     """Enable volume driven by Chromecast protocol."""
-    volume_step: float = 10
+    volume_step: int = 10
     """Volume step (1 to 100)."""
